+++ packages/veui/src/mixins/overlay.js
@@ -1,18 +1,33 @@
-import { isObject, isString } from 'lodash'
-import { normalizeClass } from '../utils/helper'
+import { normalizeClass, getClassPropDef, resolveOverlayPosition } from '../utils/helper'
+import { omit } from 'lodash'
 
 export default {
   props: {
-    overlayClass: {
-      validator (value) {
-        return isObject(value) || isString(value)
-      },
-      default: null
+    overlayClass: getClassPropDef(),
+    overlayOptions: {
+      default () {
+        return {}
+      }
+    }
+  },
+  data () {
+    return {
+      localOverlayOptions: {}
+    }
+  },
+  computed: {
+    realOverlayOptions () {
+      let options = { ...this.localOverlayOptions, ...this.overlayOptions }
+      let { position } = options
+      return omit({
+        ...options,
+        ...resolveOverlayPosition(position)
+      }, 'position')
     }
   },
   methods: {
-    mergeOverlayClass (clazz) {
-      return { ...normalizeClass(this.overlayClass), ...normalizeClass(clazz) }
+    mergeOverlayClass (klass) {
+      return { ...normalizeClass(this.overlayClass), ...normalizeClass(klass) }
     }
   }
 }